feat: add searchview and refactor exports
