+++ rs-soap-shop/src/pages/productsPage/ProductsPage.tsx
@@ -1,5 +1,5 @@
 import OurProductsCards, { items } from './cardsSection/OurProductsCards';
-import { Navigation } from './navigation/navigation';
+import { NavigationView } from './navigation/navigationView';
 import { useState } from 'react';
 import { ProductCardProps } from '../../lib/interfaces';
 import BannerPageName from '../../components/bannerPageName';
@@ -14,7 +14,7 @@ function ProductsPage() {
   return (
     <>
       <BannerPageName {...{ children: 'OUR PRODUCTS' }}></BannerPageName>
-      <Navigation changeContent={changeContent} />
+      <NavigationView changeContent={changeContent} />
       <OurProductsCards products={products} />
     </>
   );
