+++ rs-soap-shop/src/pages/productsPage/navigation/navigationView.tsx
@@ -1,15 +1,17 @@
 import { SelectCategory } from './selectCategory';
-import { SortingModule } from './sorting/sortingModule';
-import { FilterModule } from './filter/filterModule';
+import SortingView from './sorting/sortingView';
+import FilterModule from './filter/filterView';
 import { OurProductsCardsProps } from '../../../lib/interfaces';
+import SearchView from './search/searchView';
 
-export function Navigation({ changeContent }: OurProductsCardsProps) {
+export function NavigationView({ changeContent }: OurProductsCardsProps) {
   return (
     <div className='bg-accentColor dark:bg-accentDarkColor text-primaryColor'>
       <div className='flex flex-wrap justify-between items-center max-w-[1440px] py-4 px-4 mx-auto lg:px-big'>
         <SelectCategory changeContent={changeContent} />
-        <SortingModule />
+        <SortingView />
         <FilterModule />
+        <SearchView />
       </div>
     </div>
   );
