+++ rs-soap-shop/src/pages/productsPage/navigation/sorting/sortByName.tsx
@@ -2,13 +2,13 @@ import { SortABC } from '../../../../icons/sortABC';
 import { SortZYX } from '../../../../icons/sortZYX';
 import { iconClassesNormal } from '../../../../lib/constants';
 
-export function SortByName() {
+export default function SortByName() {
   return (
     <div className={'flex gap-[0.5rem]'}>
-      <div className={`${iconClassesNormal}`}>
+      <div className={`${iconClassesNormal} hover:bg-grayLColor active:scale-95 transition`}>
         <SortABC />
       </div>
-      <div className={`${iconClassesNormal}`}>
+      <div className={`${iconClassesNormal} hover:bg-grayLColor active:scale-95 transition`}>
         <SortZYX />
       </div>
     </div>
