+++ rs-soap-shop/src/pages/productsPage/navigation/sorting/sortByPrice.tsx
@@ -2,13 +2,13 @@ import { SortPriceUp } from '../../../../icons/sortPriceUp';
 import { SortPriceDown } from '../../../../icons/sortPriceDown';
 import { iconClassesNormal } from '../../../../lib/constants';
 
-export function SortByPrice() {
+export default function SortByPrice() {
   return (
     <div className={'flex gap-[0.5rem]'}>
-      <div className={`${iconClassesNormal}`}>
+      <div className={`${iconClassesNormal} hover:bg-grayLColor active:scale-95 transition`}>
         <SortPriceUp />
       </div>
-      <div className={`${iconClassesNormal}`}>
+      <div className={`${iconClassesNormal} hover:bg-grayLColor active:scale-95 transition`}>
         <SortPriceDown />
       </div>
     </div>
