--- rs-soap-shop/src/pages/productsPage/navigation/sorting/sortingModule.tsx
@@ -1,11 +1,0 @@
-import { SortByName } from './sortByName';
-import { SortByPrice } from './sortByPrice';
-
-export function SortingModule() {
-  return (
-    <div className={'flex gap-[0.5rem]'}>
-      <SortByName />
-      <SortByPrice />
-    </div>
-  );
-}
