+++ rs-soap-shop/src/pages/productsPage/navigation/sorting/sortingView.tsx
@@ -1,0 +1,11 @@
+import SortByName from './sortByName';
+import SortByPrice from './sortByPrice';
+
+export default function SortingView() {
+  return (
+    <div className={'flex gap-[0.5rem]'}>
+      <SortByName />
+      <SortByPrice />
+    </div>
+  );
+}
